Change redux to module structure #14
