+++ client/src/redux/configureStore.js
@@ -1,6 +1,6 @@
 import { applyMiddleware, compose, createStore } from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer from './reducers';
+import rootReducer from './rootReducer';
 import { routerMiddleware } from 'react-router-redux';
 
 export default function configureStore (initialState = {}, history) {
@@ -17,8 +17,8 @@ export default function configureStore (initialState = {}, history) {
   const store = createStore(rootReducer, initialState, middleware);
 
   if (module.hot) {
-    module.hot.accept('./reducers', () => {
-      const nextRootReducer = require('./reducers').default;
+    module.hot.accept('./rootReducer', () => {
+      const nextRootReducer = require('./rootReducer').default;
 
       store.replaceReducer(nextRootReducer);
     });
