--- client/src/redux/constants/scan.js
@@ -1,6 +1,0 @@
-export const SCAN_SUCCESS = 'SCAN_SUCCESS';
-export const SCAN_ERROR = 'SCAN_ERROR';
-export const SCAN_LOADING = 'SCAN_LOADING';
-
-export const CHANGE_IP = 'CHANGE_IP';
-export const CHANGE_PORT = 'CHANGE_PORT';
