+++ client/src/redux/rootReducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { routerReducer as router } from 'react-router-redux';
-import scan from './scan';
+import scan from 'redux/modules/scanner';
 
 export default combineReducers({
   router,
